// src/types/productTypes/RAMType.ts
export interface RAMData {
    modelNumber: string | null;
    name: string | null;
    brand: string | null;
    standard: string | null;
    volume: number | null;
    dimmCount: number | null;
    clockRate: number | null;
    latency: string | null;
    voltage: number | null;
    isActive: boolean;
}